log exist check

+++ web_Project/Proj/yesmem_server.c
@@ -700,7 +700,8 @@ int main(int argc, char *argv[])
 		return -1;
 	}
 
-	system("rm log.txt");	//log.txt 삭제하고 시작.
+	if (access("./log.txt", 0) == 0)
+		system("rm log.txt");	//log.txt 삭제하고 시작.
 
 	//이 부분들이 없으면 , 작업도중에 엑박남.
 	(void)signal(SIGCLD, SIG_IGN);
